pyproject.toml
[project]
name = "family-electricity-cost-share"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
